1494 - Move test cases and schemas to scripts that already exist
