+++ projects/_03_independent_cqc/_07_estimate_filled_posts_by_job_role/fargate/_03_impute.py
@@ -5,7 +5,6 @@ from projects._03_independent_cqc._07_estimate_filled_posts_by_job_role.fargate.
     create_ascwds_job_role_rolling_ratio,
     create_imputed_ascwds_job_role_counts,
 )
-from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
 
 # Set streaming chunk size for memory management - each thread (per CPU core) will load
 # in a chunk of this size.
